requirements.txt
streamlit
Pillow
numpy
torch
torchvision
st-supabase-connection
st-star-rating
matplotlib
tensorflow
tensorflow-hub